Move utils to I18NUtils
